+++ src/main/webapp/entries.js
@@ -57,6 +57,7 @@ module.exports = {
     "./resources/js/pages/projects/project-users-groups.js",
   "associated-projects":
     "./resources/js/pages/projects/associated-projects/index.js",
+  sample: "./resources/js/pages/SamplePage.jsx",
   "sample-files": "./resources/js/pages/samples/sample-files.js",
   "project-reference-files":
     "./resources/js/pages/projects/project-reference-files.js",